+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 # Version 0.2.3-SNAPSHOT
+- It is now required to supply all components at the `Engine` creation
 - `ensureCapacity(int)` for all `Bag` implementations
 - Fix: `Bag.set(int)` no longer treats the `0` value specially